Changes for Stress Testing

+++ setup/delete-kafkasources.sh
@@ -1,2 +1,4 @@
-echo "Deleting Kafka sources"
-kubectl get kafkasources.sources.knative.dev --no-headers=true | awk '{print $1}' | xargs kubectl delete kafkasources.sources.knative.dev 
+if [ -z $SKIP_CLEANUP ]; then
+  echo "Deleting Kafka sources"
+  kubectl get kafkasources.sources.knative.dev --no-headers=true | awk '{print $1}' | xargs kubectl delete kafkasources.sources.knative.dev
+fi
